Polish apache/dubbo-spring-boot-project#718 : [Samples] Introducing the samples deployed the external Servlet container
