+++ dubbo-spring-boot-actuator/README.md
@@ -201,7 +201,7 @@ Actuator endpoint `dubbo` supports Actuator Endpoints :
   "dubbo.protocol.port": "12345",
   "dubbo.registry.address": "N/A",
   "dubbo.registry.id": "my-registry",
-  "dubbo.scan.basePackages": "org.apache.dubbo.spring.boot.demo.provider.service"
+  "dubbo.scan.basePackages": "org.apache.dubbo.spring.boot.sample.provider.service"
 }
 ```
 
@@ -217,7 +217,7 @@ The structure of JSON is simple Key-Value format , the key is property name as a
 
 ```json
 {
-  "ServiceBean@org.apache.dubbo.spring.boot.demo.api.DemoService#defaultDemoService": {
+  "ServiceBean@org.apache.dubbo.spring.boot.sample.api.DemoService#defaultDemoService": {
     "accesslog": null,
     "actives": null,
     "cache": null,
@@ -233,9 +233,9 @@ The structure of JSON is simple Key-Value format , the key is property name as a
     "filter": "",
     "generic": "false",
     "group": null,
-    "id": "org.apache.dubbo.spring.boot.demo.api.DemoService",
-    "interface": "org.apache.dubbo.spring.boot.demo.api.DemoService",
-    "interfaceClass": "org.apache.dubbo.spring.boot.demo.api.DemoService",
+    "id": "org.apache.dubbo.spring.boot.sample.api.DemoService",
+    "interface": "org.apache.dubbo.spring.boot.sample.api.DemoService",
+    "interfaceClass": "org.apache.dubbo.spring.boot.sample.api.DemoService",
     "layer": null,
     "listener": "",
     "loadbalance": null,
@@ -245,7 +245,7 @@ The structure of JSON is simple Key-Value format , the key is property name as a
     "onconnect": null,
     "ondisconnect": null,
     "owner": null,
-    "path": "org.apache.dubbo.spring.boot.demo.api.DemoService",
+    "path": "org.apache.dubbo.spring.boot.sample.api.DemoService",
     "proxy": null,
     "retries": null,
     "scope": null,
@@ -254,7 +254,7 @@ The structure of JSON is simple Key-Value format , the key is property name as a
     "timeout": null,
     "token": null,
     "unexported": false,
-    "uniqueServiceName": "org.apache.dubbo.spring.boot.demo.api.DemoService:1.0.0",
+    "uniqueServiceName": "org.apache.dubbo.spring.boot.sample.api.DemoService:1.0.0",
     "validation": null,
     "version": "1.0.0",
     "warmup": null,
@@ -274,7 +274,7 @@ The key is the Bean name of `ServiceBean` , `ServiceBean`'s properties compose v
 
 ```json
 {
-  "private org.apache.dubbo.spring.boot.demo.api.DemoService org.apache.dubbo.spring.boot.demo.consumer.controller.DemoConsumerController.demoService": {
+  "private org.apache.dubbo.spring.boot.sample.api.DemoService org.apache.dubbo.spring.boot.sample.consumer.controller.DemoConsumerController.demoService": {
     "actives": null,
     "cache": null,
     "callbacks": null,
@@ -285,9 +285,9 @@ The key is the Bean name of `ServiceBean` , `ServiceBean`'s properties compose v
     "filter": "",
     "generic": null,
     "group": null,
-    "id": "org.apache.dubbo.spring.boot.demo.api.DemoService",
-    "interface": "org.apache.dubbo.spring.boot.demo.api.DemoService",
-    "interfaceClass": "org.apache.dubbo.spring.boot.demo.api.DemoService",
+    "id": "org.apache.dubbo.spring.boot.sample.api.DemoService",
+    "interface": "org.apache.dubbo.spring.boot.sample.api.DemoService",
+    "interfaceClass": "org.apache.dubbo.spring.boot.sample.api.DemoService",
     "layer": null,
     "lazy": null,
     "listener": "",
@@ -295,7 +295,7 @@ The key is the Bean name of `ServiceBean` , `ServiceBean`'s properties compose v
     "local": null,
     "merger": null,
     "mock": null,
-    "objectType": "org.apache.dubbo.spring.boot.demo.api.DemoService",
+    "objectType": "org.apache.dubbo.spring.boot.sample.api.DemoService",
     "onconnect": null,
     "ondisconnect": null,
     "owner": null,
@@ -310,7 +310,7 @@ The key is the Bean name of `ServiceBean` , `ServiceBean`'s properties compose v
     "stub": null,
     "stubevent": null,
     "timeout": null,
-    "uniqueServiceName": "org.apache.dubbo.spring.boot.demo.api.DemoService:1.0.0",
+    "uniqueServiceName": "org.apache.dubbo.spring.boot.sample.api.DemoService:1.0.0",
     "url": "dubbo://localhost:12345",
     "validation": null,
     "version": "1.0.0",
