--- dubbo-spring-boot-samples/auto-configure-samples/provider-sample/src/main/java/org/apache/dubbo/spring/boot/demo/provider/service/DefaultDemoService.java
@@ -1,62 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.dubbo.spring.boot.demo.provider.service;
-
-import org.apache.dubbo.config.annotation.DubboService;
-import org.apache.dubbo.spring.boot.demo.consumer.DemoService;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-
-import java.util.Random;
-
-/**
- * Default {@link DemoService}
- *
- * @see DemoService
- * @since 2.7.0
- */
-@DubboService(version = "1.0.0")
-public class DefaultDemoService implements DemoService {
-
-    private final Logger logger = LoggerFactory.getLogger(getClass());
-
-    private final Random costTimeRandom = new Random();
-
-    /**
-     * The default value of ${dubbo.application.name} is ${spring.application.name}
-     */
-    @Value("${dubbo.application.name}")
-    private String serviceName;
-
-    @Override
-    public String sayHello(String name) {
-        await();
-        return String.format("[%s] : Hello, %s", serviceName, name);
-    }
-
-    private void await() {
-        try {
-            long timeInMillisToWait = costTimeRandom.nextInt(500);
-            Thread.sleep(timeInMillisToWait);
-            logger.info("execution time : " + timeInMillisToWait + " ms.");
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
