+++ dubbo-spring-boot-samples/auto-configure-samples/provider-sample/src/main/java/org/apache/dubbo/spring/boot/sample/provider/bootstrap/DubboAutoConfigurationProviderBootstrap.java
@@ -14,30 +14,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.spring.boot.sample.provider.service;
+package org.apache.dubbo.spring.boot.sample.provider.bootstrap;
 
-import org.apache.dubbo.config.annotation.DubboService;
-import org.apache.dubbo.spring.boot.sample.consumer.DemoService;
+import org.apache.dubbo.spring.boot.sample.provider.service.DefaultDemoService;
 
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 
 /**
- * Default {@link DemoService}
+ * Dubbo Auto-Configuration Provider Bootstrap
  *
- * @see DemoService
+ * @see DefaultDemoService
  * @since 2.7.0
  */
-@DubboService(version = "${demo.service.version}")
-public class DefaultDemoService implements DemoService {
+@EnableAutoConfiguration
+public class DubboAutoConfigurationProviderBootstrap {
 
-    /**
-     * The default value of ${dubbo.application.name} is ${spring.application.name}
-     */
-    @Value("${dubbo.application.name}")
-    private String serviceName;
-
-    @Override
-    public String sayHello(String name) {
-        return String.format("[%s] : Hello, %s", serviceName, name);
+    public static void main(String[] args) {
+//        new SpringApplicationBuilder(DubboAutoConfigurationProviderBootstrap.class)
+//                .run(args);
+        SpringApplication.run(DubboAutoConfigurationProviderBootstrap.class,args);
     }
 }
