+++ dubbo-spring-boot-samples/auto-configure-samples/provider-sample/src/main/java/org/apache/dubbo/spring/boot/sample/provider/service/DefaultDemoService.java
@@ -19,7 +19,11 @@ package org.apache.dubbo.spring.boot.sample.provider.service;
 import org.apache.dubbo.config.annotation.DubboService;
 import org.apache.dubbo.spring.boot.sample.consumer.DemoService;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
+
+import java.util.Random;
 
 /**
  * Default {@link DemoService}
@@ -27,8 +31,12 @@ import org.springframework.beans.factory.annotation.Value;
  * @see DemoService
  * @since 2.7.0
  */
-@DubboService(version = "${demo.service.version}")
+@DubboService(version = "1.0.0")
 public class DefaultDemoService implements DemoService {
+
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    private final Random costTimeRandom = new Random();
 
     /**
      * The default value of ${dubbo.application.name} is ${spring.application.name}
@@ -38,6 +46,17 @@ public class DefaultDemoService implements DemoService {
 
     @Override
     public String sayHello(String name) {
+        await();
         return String.format("[%s] : Hello, %s", serviceName, name);
+    }
+
+    private void await() {
+        try {
+            long timeInMillisToWait = costTimeRandom.nextInt(500);
+            Thread.sleep(timeInMillisToWait);
+            logger.info("execution time : " + timeInMillisToWait + " ms.");
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
