+++ dubbo-spring-boot-samples/registry-samples/zookeeper-samples/consumer-sample/src/main/java/org/apache/dubbo/spring/boot/sample/consumer/bootstrap/DubboRegistryZooKeeperConsumerBootstrap.java
@@ -30,7 +30,7 @@ import org.springframework.context.annotation.Bean;
  * Dubbo Registry ZooKeeper Consumer Bootstrap
  */
 @EnableAutoConfiguration
-public class DubboZooKeeperServiceIntrospectionConsumerBootstrap {
+public class DubboRegistryZooKeeperConsumerBootstrap {
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
@@ -38,7 +38,7 @@ public class DubboZooKeeperServiceIntrospectionConsumerBootstrap {
     private DemoService demoService;
 
     public static void main(String[] args) {
-        SpringApplication.run(DubboZooKeeperServiceIntrospectionConsumerBootstrap.class).close();
+        SpringApplication.run(DubboRegistryZooKeeperConsumerBootstrap.class).close();
     }
 
     @Bean
