+++ dubbo-spring-boot-samples/servlet-container-samples/provider-sample/src/main/resources/application.properties
@@ -1,15 +1,13 @@
 # Spring boot application
-spring.application.name=dubbo-externalized-configuration-provider-sample
+spring.application.name=dubbo-spring-boot-servlet-container-provider-sample
 # Base packages to scan Dubbo Component: @org.apache.dubbo.config.annotation.Service
 dubbo.scan.base-packages=org.apache.dubbo.spring.boot.sample.provider.service
 
 # Dubbo Application
-## The default value of dubbo.application.name is ${spring.application.name}
-## dubbo.application.name=${spring.application.name}
 
 # Dubbo Protocol
 dubbo.protocol.name=dubbo
-dubbo.protocol.port=12345
+dubbo.protocol.port=23456
 
 ## Dubbo Registry
 dubbo.registry.address=N/A
